feat: Caching api response and application optimization's  work
Using indexedDB to cache the result of the REST Api call. The call is initialized on the UI and intercepted by the service worker, if the data is already in the local db is returned immediately and then the network call is executed, if the data is not ready yet in the db it will fetched from the network.

For the image optimization was added a library which will fetch and load the images before to be visualized on the viewport instead of the default approach of fetch all the image element on the DOM tree even when they are not been visualized.

Resolves: #15, #16