feat: Network status, new server, review form and reviews
Added  snackbar component to the restaurant.html page to notify network status changes. Closes #9.

Added new server used on the stage 3. Closes #18.

Added logic to fetch the restaurant reviews. Closes #21.

Added Material Design form to submit a review. Closes #22.

Added logic to submit new reviews. If the app is offline, when a review is submitted, the UI is updated and the review is saved in IndexDB.  Closes #23.

Resolves: #9,#18,#23,#22,#21
See also: #19,#20